Add files via upload
I unleashed my mind in this project from tools I had at my disoposal and created a assumption that counties with over a certain population percentage of Caucasian or a total population less than the county average would be called for Trump. Using this belief on census data, I created a decision tree that when boosted, revealed the true decisive variable in the election. This at the time shocked me and was incredibly pleased. This part can be found at the bottom of the pdf. I also wrote details on why I was doing something. Overall, most fun project I have ever done. Me and my partner received a 98.5% on this project with bayesian approach being hailed as "Highly Creative"